services/s3/requirements.txt
python-dotenv
awscli
boto3